Time: 56 ms (58.75%), Space: 89.8 MB (31.19%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.cpp
@@ -5,6 +5,6 @@ public:
         vector<int> save(all);
         merge(nums1.begin(), nums1.end(), nums2.begin(), nums2.end(), save.begin());
         
-        return all & 1 ? save[all / 2] : ((double)save[all / 2 - 1] + save[all / 2]) / 2;
+        return all & 1 ? save[all / 2] : (save[all / 2 - 1] + save[all / 2]) / 2.0;
     }
 };
